fix: Use clip to make ListItem rounded
